requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-filter==23.5
django-mptt==0.15.0
django-axes==6.1.1
drf-spectacular==0.27.0
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
django-environ==0.11.2